zng_wgt_slider

Trait SelectorValue

pub trait SelectorValue: VarValue {
    // Required method
    fn to_selector(value: Var<Self>, min: Self, max: Self) -> Selector;
}

Represents a type that can auto implement a Selector.

Implementing

This trait is implemented for all primitive type and Zng layout types, if a type does not you can declare custom conversions using Selector::value.

Required Methods

fn to_selector(value: Var<Self>, min: Self, max: Self) -> Selector

Make the selector.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.