Crate zng_wgt_slider

Widget for selecting a value or range by dragging a selector thumb.

Crate

This crate is part of the zng project.

Modules

thumb

Slider thumb widget.

Structs

DefaultStyle

W Default slider style.

Selector

Defines the values and ranges selected by a slider.

Slider

W Value selector from a range of values.

SliderTrack

W Slider track container widget.

ThumbArgs

Arguments for a slider thumb widget generator.

ThumbValue

Represents a selector thumb in a slider.

Enums

SliderDirection

Slider orientation and direction.

Statics

SELECTOR

Contextual Selector.

SLIDER_DIRECTION_VAR

Orientation and direction of the parent slider.

STYLE_FN_VAR

Contextual style variable.

THUMB_FN_VAR

Contextual thumb function.

Traits

SelectorValue

Represents a type that can auto implement a Selector.

WidgetInfoExt

Slider extension methods for widget info.

Functions

selector

P Sets the slider selector that defines the values, ranges that are selected.

style_fn

P Extends or replaces the widget style.

thumb_fn

P Widget function that converts ThumbArgs to widgets.